# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Simple User Post",
    "description": "Aplikasi Untuk Posting",
    "termsOfService": "http://didi.com/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.didi.com/support",
      "email": "support@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    },
    {
      "url": "http://production.com",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/post":{
      "post": {
        "tags": ["Posts"],
        "summary": "Create A New Post",
        "operationId": "createPost",
        "requestBody": {
          "description": "Buat Post Baru",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Create_Post_Body"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Create_Post_Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "sucessful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref":"#/components/schemas/Post"
                    }
                  }
                 
                }
              }
            }
          },
          "500": {
            "description": "error operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get":{
        "tags": ["Posts"],
        "summary": "Get All Posts",
        "operationId": "getAllPosts",
        "description": "Get All Post From Database",
        "responses": {
          "200": {
            "description": "Successfully Get Posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Posts Successfully Found"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Post_User"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Post Data Is Empty",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": " Post Data Is Empty"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Some Error Message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "required": ["name", "email", "role"],
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date"
          },
          "updated_at": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "Post": {
        "required": ["title", "content", "user_uuid"],
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "user_uuid": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date"
          },
          "updated_at": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "Post_User": {
        "required": ["title", "content", "user_uuid"],
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "user_uuid": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date"
          },
          "updated_at": {
            "type": "string",
            "format": "date"
          },
          "user": {
            "type": "array",
            "items": {
              "$ref":"#/components/schemas/User"
            }
          }
        }
      },
      "Create_Post_Body": {
        "required": ["user_uuid", "title", "content"],
        "type": "object",
        "properties": {
          "user_uuid": {
            "type": "string",
            "example": "3dcf17a3-2a63-4cf4-94c7-9c21283629b2"
          },
          "title": {
            "type": "string",
            "example": "some title"
          },
          "content": {
            "type": "string",
            "example": "some content"
          }
        }
      },
...
    }
  }
}